requirements.txt
asgiref==3.8.1
channels==4.1.0
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
Faker==25.8.0
joblib==1.4.2
numpy==1.26.4
pillow==10.4.0
python-dateutil==2.9.0.post0
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
sqlparse==0.5.0
threadpoolctl==3.5.0
tzdata==2024.1
